refactor(runtime): split builtins_collection.rs into cohesive modules (§7-8)
Split the 2587-line builtins_collection.rs (over the 500-line limit)
into seven cohesive modules, all under 500 lines:

- builtins_collection.rs (392)          - module-level free fns
                                          (val/minmax-pub/first-result) +
                                          constructors (set/bag/mix/hash/
                                          pair/junction/keys/values/kv/pairs)
- builtins_collection_extrema.rs (313)  - min/max + extrema helpers
- builtins_collection_minmax.rs (303)   - minmax + adverb application
- builtins_collection_listops.rs (356)  - shift/pop/join/sort/unique/...
- builtins_collection_mapgrep.rs (438)  - map/grep/snip/first
- builtins_collection_classify.rs (339) - classify/categorize
- builtins_collection_deepmap.rs (460)  - cross/roundrobin/duckmap/deepmap

All methods are impl Interpreter, so the split is visibility-only. The
module-level free fns (collect_minmax_candidates_pub,
make_inclusive_range_pub, format_first_result, builtin_val) stay in the
main module to preserve the crate::runtime::builtins_collection::* paths
used by callers. A few cross-module helpers (order_to_int,
find_var_by_identity) are bumped to pub(super), and mapgrep imports
format_first_result. The fn multiset is byte-identical. No behavior change.
