+++ src/runtime/builtins_collection.rs
@@ -31,6 +31,56 @@ pub(super) fn format_first_result(
     }
 }
 
+/// Raku `val()` builtin: convert a string into an allomorphic type.
+pub(crate) fn builtin_val(args: &[Value]) -> Value {
+    let arg = match args.first() {
+        Some(v) => v,
+        None => return Value::Nil,
+    };
+    // val() on non-Str types (List, Slip, Array) returns the value unchanged.
+    match arg {
+        Value::Array(..) | Value::Seq(_) | Value::Slip(_) => return arg.clone(),
+        _ => {}
+    }
+    let original = arg.to_string_value();
+    let word = original.trim();
+
+    fn make_allomorphic(val: Value, original: &str) -> Value {
+        let mut mixins = StdHashMap::new();
+        // Store the original string (with whitespace) as the Str component
+        mixins.insert("Str".to_string(), Value::str(original.to_string()));
+        Value::mixin(val, mixins)
+    }
+
+    // Try Unicode vulgar fractions (single character like ½, ⅓, ¼, etc.)
+    if let Some(rat_val) = try_parse_unicode_fraction(word) {
+        return make_allomorphic(rat_val, &original);
+    }
+
+    // Use the comprehensive Raku numeric string parser
+    if let Some(numeric) = crate::runtime::str_numeric::parse_raku_str_to_numeric(word) {
+        return make_allomorphic(numeric, &original);
+    }
+
+    // Plain string (not parseable as a number)
+    Value::str(original.to_string())
+}
+
+/// Try to parse a single Unicode vulgar fraction character (½, ⅓, ¼, etc.)
+fn try_parse_unicode_fraction(s: &str) -> Option<Value> {
+    let mut chars = s.chars();
+    let ch = chars.next()?;
+    // Must be exactly one character
+    if chars.next().is_some() {
+        return None;
+    }
+    let (n, d) = crate::builtins::unicode::unicode_rat_value(ch)?;
+    if d == 0 {
+        return None;
+    }
+    Some(crate::value::make_rat(n, d))
+}
+
 impl Interpreter {
     pub(super) fn builtin_end(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
         if args.len() != 1 {
@@ -338,2250 +388,5 @@ impl Interpreter {
         let target = args.first().cloned().unwrap_or(Value::Nil);
         crate::builtins::native_method_0arg(&target, crate::symbol::Symbol::intern(method))
             .unwrap_or_else(|| Ok(Value::array(Vec::new())))
-    }
-
-    // builtin_abs / builtin_sign removed (Slice 6.3 / Category A dedup): native
-    // `abs` / `sign` in src/builtins/functions.rs are authoritative (reached via
-    // call_function_fallback -> native_function).
-}
-
-/// Raku `val()` builtin: convert a string into an allomorphic type.
-pub(crate) fn builtin_val(args: &[Value]) -> Value {
-    let arg = match args.first() {
-        Some(v) => v,
-        None => return Value::Nil,
-    };
-    // val() on non-Str types (List, Slip, Array) returns the value unchanged.
-    match arg {
-        Value::Array(..) | Value::Seq(_) | Value::Slip(_) => return arg.clone(),
-        _ => {}
-    }
-    let original = arg.to_string_value();
-    let word = original.trim();
-
-    fn make_allomorphic(val: Value, original: &str) -> Value {
-        let mut mixins = StdHashMap::new();
-        // Store the original string (with whitespace) as the Str component
-        mixins.insert("Str".to_string(), Value::str(original.to_string()));
-        Value::mixin(val, mixins)
-    }
-
-    // Try Unicode vulgar fractions (single character like ½, ⅓, ¼, etc.)
-    if let Some(rat_val) = try_parse_unicode_fraction(word) {
-        return make_allomorphic(rat_val, &original);
-    }
-
-    // Use the comprehensive Raku numeric string parser
-    if let Some(numeric) = crate::runtime::str_numeric::parse_raku_str_to_numeric(word) {
-        return make_allomorphic(numeric, &original);
-    }
-
-    // Plain string (not parseable as a number)
-    Value::str(original.to_string())
-}
-
-/// Try to parse a single Unicode vulgar fraction character (½, ⅓, ¼, etc.)
-fn try_parse_unicode_fraction(s: &str) -> Option<Value> {
-    let mut chars = s.chars();
-    let ch = chars.next()?;
-    // Must be exactly one character
-    if chars.next().is_some() {
-        return None;
-    }
-    let (n, d) = crate::builtins::unicode::unicode_rat_value(ch)?;
-    if d == 0 {
-        return None;
-    }
-    Some(crate::value::make_rat(n, d))
-}
-
-impl Interpreter {
-    fn failure_exception_from_value(value: &Value) -> Option<Value> {
-        match value {
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "Failure" => attributes.as_map().get("exception").cloned(),
-            Value::Mixin(inner, mixins) => {
-                if let Some(mixed) = mixins.get("Failure")
-                    && let Some(ex) = Self::failure_exception_from_value(mixed)
-                {
-                    return Some(ex);
-                }
-                Self::failure_exception_from_value(inner)
-            }
-            _ => None,
-        }
-    }
-
-    fn is_failure_like(value: &Value) -> bool {
-        Self::failure_exception_from_value(value).is_some()
-    }
-
-    /// Determine the arity of a callable for min/max/minmax by-block dispatch.
-    /// Returns 1 for sort-key extractors, 2 for comparators.
-    pub(crate) fn extrema_callable_arity(&self, callable: &Value) -> usize {
-        match callable {
-            Value::Sub(data) => {
-                if data.params.is_empty() {
-                    1 // implicit $_ → arity 1
-                } else {
-                    data.params.len()
-                }
-            }
-            Value::WeakSub(weak) => {
-                if let Some(data) = weak.upgrade() {
-                    if data.params.is_empty() {
-                        1
-                    } else {
-                        data.params.len()
-                    }
-                } else {
-                    2
-                }
-            }
-            Value::Routine { .. } => {
-                let (params, param_defs) = self.callable_signature(callable);
-                if !param_defs.is_empty() {
-                    let positional_count = param_defs
-                        .iter()
-                        .filter(|pd| !pd.named && !pd.slurpy && !pd.double_slurpy)
-                        .count();
-                    if positional_count > 0 {
-                        return positional_count;
-                    }
-                }
-                if params.is_empty() { 1 } else { params.len() }
-            }
-            _ => 2, // default to comparator
-        }
-    }
-
-    fn extrema_from_values_by(
-        &mut self,
-        args: &[Value],
-        want_max: bool,
-        by: Option<&Value>,
-    ) -> Result<Value, RuntimeError> {
-        // Determine if by-block is arity-1 (sort key) or arity-2 (comparator)
-        // before borrowing self mutably for the block-call closure.
-        let by_arity = by.map(|b| self.extrema_callable_arity(b));
-        Self::extrema_from_values_generic(args, want_max, by, by_arity, |c, a| {
-            self.call_sub_value(c.clone(), a, true)
-        })
-    }
-
-    /// Engine-agnostic min/max fold over a flat list of values (the genuine
-    /// Category-B fork is the `:by` block, which each engine supplies through
-    /// `call`). Shared by the interpreter (`extrema_from_values_by`) and the VM
-    /// (`try_native_extrema`) so the flatten / failure / `Any`-filter / fold
-    /// logic exists once. `by_arity` is the resolved arity of `by`
-    /// (`extrema_callable_arity`); 1 = sort-key mapper, otherwise a comparator.
-    pub(crate) fn extrema_from_values_generic<F>(
-        args: &[Value],
-        want_max: bool,
-        by: Option<&Value>,
-        by_arity: Option<usize>,
-        mut call: F,
-    ) -> Result<Value, RuntimeError>
-    where
-        F: FnMut(&Value, Vec<Value>) -> Result<Value, RuntimeError>,
-    {
-        if args.is_empty() {
-            return Ok(Value::Nil);
-        }
-
-        // Flatten: if a single array/seq/list arg is provided, use its items
-        let expanded: Vec<Value> = if args.len() == 1 {
-            if let Some(items) = args[0].as_list_items() {
-                items.to_vec()
-            } else {
-                args.to_vec()
-            }
-        } else {
-            args.to_vec()
-        };
-
-        if expanded.is_empty() {
-            // Empty list: return Inf for min, -Inf for max
-            return Ok(if want_max {
-                Value::Num(f64::NEG_INFINITY)
-            } else {
-                Value::Num(f64::INFINITY)
-            });
-        }
-
-        let failures: Vec<Value> = expanded
-            .iter()
-            .filter(|v| Self::is_failure_like(v))
-            .cloned()
-            .collect();
-        if failures.len() >= 2
-            && let Some(ex) = Self::failure_exception_from_value(&failures[0])
-        {
-            let mut err = RuntimeError::new(ex.to_string_value());
-            err.exception = Some(Box::new(ex));
-            return Err(err);
-        }
-        if failures.len() == 1 {
-            return Ok(failures[0].clone());
-        }
-
-        let mut filtered: Vec<Value> = expanded
-            .iter()
-            .filter(|v| !matches!(v, Value::Package(name) if name == "Any"))
-            .cloned()
-            .collect();
-        if filtered.is_empty() {
-            return Ok(args[0].clone());
-        }
-
-        let mut best = filtered.remove(0);
-        for value in filtered {
-            let cmp = if let Some(by_block) = by {
-                if by_arity == Some(1) {
-                    // Arity-1: use as sort key extractor
-                    let key_a = call(by_block, vec![value.clone()])?;
-                    let key_b = call(by_block, vec![best.clone()])?;
-                    crate::runtime::compare_values(&key_a, &key_b)
-                } else {
-                    // Arity-2: use as comparator ($^a, $^b) => Order
-                    let result = call(by_block, vec![value.clone(), best.clone()])?;
-                    Self::order_to_int(&result)
-                }
-            } else {
-                crate::runtime::compare_values(&value, &best)
-            };
-            if (want_max && cmp > 0) || (!want_max && cmp < 0) {
-                best = value;
-            }
-        }
-        Ok(best)
-    }
-
-    fn extrema_from_hash(
-        &mut self,
-        map: &std::sync::Arc<crate::value::HashData>,
-        by: Option<Value>,
-        want_max: bool,
-    ) -> Result<Value, RuntimeError> {
-        let mut best_pair: Option<Value> = None;
-        let mut best_key: Option<Value> = None;
-
-        for (k, v) in map.as_ref().iter() {
-            // Object hashes store `.WHICH` string keys; expose the original
-            // typed key both in the resulting Pair and in the default ordering.
-            let pair = map.typed_pair(k, v.clone());
-            let key = if let Some(by_callable) = &by {
-                match self.call_sub_value(by_callable.clone(), vec![pair.clone()], true) {
-                    Ok(value) => value,
-                    Err(_) => v.clone(),
-                }
-            } else {
-                map.typed_key(k)
-            };
-
-            let replace = if let Some(current_key) = &best_key {
-                let cmp = crate::runtime::compare_values(&key, current_key);
-                (want_max && cmp > 0) || (!want_max && cmp < 0)
-            } else {
-                true
-            };
-            if replace {
-                best_pair = Some(pair);
-                best_key = Some(key);
-            }
-        }
-
-        Ok(best_pair.unwrap_or(Value::Nil))
-    }
-
-    /// Extract named adverbs (:k, :v, :kv, :p) from args for min/max.
-    pub(crate) fn extract_extrema_adverbs(args: &[Value]) -> (Option<String>, Vec<Value>) {
-        let mut adverb = None;
-        let mut positional = Vec::new();
-        let mut by_found = false;
-        for arg in args {
-            match arg {
-                Value::Pair(name, _) if name == "by" => {
-                    by_found = true;
-                    positional.push(arg.clone()); // keep by pair for later extraction
-                }
-                Value::ValuePair(key, _) if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by") =>
-                {
-                    by_found = true;
-                    positional.push(arg.clone());
-                }
-                Value::Pair(name, value)
-                    if matches!(name.as_str(), "k" | "v" | "kv" | "p")
-                        && matches!(value.as_ref(), Value::Bool(true)) =>
-                {
-                    adverb = Some(name.clone());
-                }
-                _ => positional.push(arg.clone()),
-            }
-        }
-        let _ = by_found;
-        (adverb, positional)
-    }
-
-    pub(super) fn builtin_min(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let (adverb, filtered_args) = Self::extract_extrema_adverbs(args);
-        let result = self.builtin_min_inner(&filtered_args)?;
-        self.apply_extrema_adverb(&filtered_args, result, false, adverb.as_deref())
-    }
-
-    fn builtin_min_inner(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let by = args.iter().find_map(|arg| match arg {
-            Value::Pair(name, value) if name == "by" => Some((**value).clone()),
-            Value::ValuePair(key, value)
-                if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by") =>
-            {
-                Some((**value).clone())
-            }
-            _ => None,
-        });
-        let positional: Vec<Value> = args
-            .iter()
-            .filter(|arg| {
-                !matches!(arg, Value::Pair(name, _) if name == "by")
-                    && !matches!(arg, Value::ValuePair(key, _) if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by"))
-            })
-            .cloned()
-            .collect();
-
-        if positional.len() == 1
-            && let Value::Hash(map) = &positional[0]
-        {
-            return self.extrema_from_hash(map, by, false);
-        }
-        if positional.len() == 1
-            && let Value::Instance { class_name, .. } = &positional[0]
-            && class_name == "Hash"
-        {
-            let method_args = by.map_or_else(Vec::new, |v| vec![v]);
-            return self.call_method_with_values(positional[0].clone(), "min", method_args);
-        }
-        self.extrema_from_values_by(&positional, false, by.as_ref())
-    }
-
-    pub(super) fn builtin_max(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let (adverb, filtered_args) = Self::extract_extrema_adverbs(args);
-        let result = self.builtin_max_inner(&filtered_args)?;
-        self.apply_extrema_adverb(&filtered_args, result, true, adverb.as_deref())
-    }
-
-    fn builtin_max_inner(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let by = args.iter().find_map(|arg| match arg {
-            Value::Pair(name, value) if name == "by" => Some((**value).clone()),
-            Value::ValuePair(key, value)
-                if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by") =>
-            {
-                Some((**value).clone())
-            }
-            _ => None,
-        });
-        let positional: Vec<Value> = args
-            .iter()
-            .filter(|arg| {
-                !matches!(arg, Value::Pair(name, _) if name == "by")
-                    && !matches!(arg, Value::ValuePair(key, _) if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by"))
-            })
-            .cloned()
-            .collect();
-
-        if positional.len() == 1
-            && let Value::Hash(map) = &positional[0]
-        {
-            return self.extrema_from_hash(map, by, true);
-        }
-        if positional.len() == 1
-            && let Value::Instance { class_name, .. } = &positional[0]
-            && class_name == "Hash"
-        {
-            let method_args = by.map_or_else(Vec::new, |v| vec![v]);
-            return self.call_method_with_values(positional[0].clone(), "max", method_args);
-        }
-        self.extrema_from_values_by(&positional, true, by.as_ref())
-    }
-
-    /// Apply :k, :v, :kv, :p adverbs to min/max result.
-    /// These return index/value/both/pair information about the extremum.
-    pub(crate) fn apply_extrema_adverb(
-        &self,
-        args: &[Value],
-        result: Value,
-        _want_max: bool,
-        adverb: Option<&str>,
-    ) -> Result<Value, RuntimeError> {
-        let adverb = match adverb {
-            Some(a) => a,
-            None => return Ok(result),
-        };
-
-        // Get the list items from the first positional arg
-        let positional: Vec<Value> = args
-            .iter()
-            .filter(|arg| {
-                !matches!(arg, Value::Pair(_, _)) && !matches!(arg, Value::ValuePair(_, _))
-            })
-            .cloned()
-            .collect();
-
-        // Flatten list items
-        let items: Vec<Value> = if positional.len() == 1 {
-            if let Some(list) = positional[0].as_list_items() {
-                list.to_vec()
-            } else {
-                positional
-            }
-        } else {
-            positional
-        };
-
-        if items.is_empty() {
-            return Ok(Value::Seq(std::sync::Arc::new(vec![])));
-        }
-
-        // Find all indices matching the extremum
-        let mut matching_indices = Vec::new();
-        for (i, item) in items.iter().enumerate() {
-            let cmp = crate::runtime::compare_values(item, &result);
-            if cmp == 0 {
-                matching_indices.push(i);
-            }
-        }
-
-        match adverb {
-            "k" => {
-                let keys: Vec<Value> = matching_indices
-                    .iter()
-                    .map(|&i| Value::Int(i as i64))
-                    .collect();
-                Ok(Value::Seq(std::sync::Arc::new(keys)))
-            }
-            "v" => {
-                let vals: Vec<Value> = matching_indices.iter().map(|&i| items[i].clone()).collect();
-                Ok(Value::Seq(std::sync::Arc::new(vals)))
-            }
-            "kv" => {
-                let mut kvs = Vec::new();
-                for &i in &matching_indices {
-                    kvs.push(Value::Int(i as i64));
-                    kvs.push(items[i].clone());
-                }
-                Ok(Value::Seq(std::sync::Arc::new(kvs)))
-            }
-            "p" => {
-                let pairs: Vec<Value> = matching_indices
-                    .iter()
-                    .map(|&i| {
-                        Value::ValuePair(Box::new(Value::Int(i as i64)), Box::new(items[i].clone()))
-                    })
-                    .collect();
-                Ok(Value::Seq(std::sync::Arc::new(pairs)))
-            }
-            _ => Ok(result),
-        }
-    }
-
-    pub(crate) fn collect_minmax_candidates(value: &Value, out: &mut Vec<Value>) {
-        match value {
-            // Unwrap Scalar containers
-            Value::Scalar(inner) => Self::collect_minmax_candidates(inner, out),
-            // Unwrap first-class element containers (`:=`-bound / grep rw alias).
-            Value::ContainerRef(cell) => {
-                Self::collect_minmax_candidates(&cell.lock().unwrap().clone(), out)
-            }
-            Value::Range(a, b)
-            | Value::RangeExcl(a, b)
-            | Value::RangeExclStart(a, b)
-            | Value::RangeExclBoth(a, b) => {
-                out.push(Value::Int(*a));
-                out.push(Value::Int(*b));
-            }
-            Value::GenericRange { start, end, .. } => {
-                // Skip Inf..-Inf (identity for minmax combine from empty lists)
-                let is_start_inf = matches!(start.as_ref(), Value::Num(n) if n.is_infinite() && n.is_sign_positive());
-                let is_end_neg_inf = matches!(end.as_ref(), Value::Num(n) if n.is_infinite() && n.is_sign_negative());
-                if is_start_inf && is_end_neg_inf {
-                    return; // skip identity range
-                }
-                out.push((**start).clone());
-                out.push((**end).clone());
-            }
-            _ if value.as_list_items().is_some() => {
-                // Recursively collect from list items so itemized sub-ranges
-                // within a tuple get their endpoints extracted
-                for item in value.as_list_items().unwrap().iter() {
-                    Self::collect_minmax_candidates(item, out);
-                }
-            }
-            other => out.push(other.clone()),
-        }
-    }
-
-    pub(crate) fn make_inclusive_range_value(left: Value, right: Value) -> Value {
-        match (&left, &right) {
-            (Value::Int(a), Value::Int(b)) => Value::Range(*a, *b),
-            (Value::Int(a), Value::Num(b)) if b.is_infinite() && b.is_sign_positive() => {
-                Value::Range(*a, i64::MAX)
-            }
-            (Value::Int(a), Value::Whatever) => Value::Range(*a, i64::MAX),
-            (Value::Num(a), Value::Int(b)) if a.is_infinite() && a.is_sign_negative() => {
-                Value::Range(i64::MIN, *b)
-            }
-            (Value::Str(a), Value::Str(b)) => Value::GenericRange {
-                start: Arc::new(Value::Str(a.clone())),
-                end: Arc::new(Value::Str(b.clone())),
-                excl_start: false,
-                excl_end: false,
-            },
-            (l, r) if l.is_numeric() && r.is_numeric() => Value::GenericRange {
-                start: Arc::new(l.clone()),
-                end: Arc::new(r.clone()),
-                excl_start: false,
-                excl_end: false,
-            },
-            (Value::Str(a), r) if r.is_numeric() => Value::GenericRange {
-                start: Arc::new(Value::Str(a.clone())),
-                end: Arc::new(r.clone()),
-                excl_start: false,
-                excl_end: false,
-            },
-            (l, Value::Str(b)) if l.is_numeric() => Value::GenericRange {
-                start: Arc::new(l.clone()),
-                end: Arc::new(Value::Str(b.clone())),
-                excl_start: false,
-                excl_end: false,
-            },
-            (_, Value::Sub(_)) | (Value::Sub(_), _) => Value::GenericRange {
-                start: Arc::new(left),
-                end: Arc::new(right),
-                excl_start: false,
-                excl_end: false,
-            },
-            _ => Value::Nil,
-        }
-    }
-
-    pub(super) fn builtin_minmax(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let by = args.iter().find_map(|arg| match arg {
-            Value::Pair(name, value) if name == "by" => Some((**value).clone()),
-            Value::ValuePair(key, value)
-                if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by") =>
-            {
-                Some((**value).clone())
-            }
-            _ => None,
-        });
-        let positional: Vec<Value> = args
-            .iter()
-            .filter(|arg| {
-                !matches!(arg, Value::Pair(name, _) if name == "by")
-                    && !matches!(arg, Value::ValuePair(key, _) if matches!(key.as_ref(), Value::Str(name) if name.as_str() == "by"))
-            })
-            .cloned()
-            .collect();
-
-        // Resolve by-block arity before borrowing self mutably for the closure.
-        let by_arity = by.as_ref().map(|b| self.extrema_callable_arity(b));
-        Self::minmax_from_values_generic(&positional, by.as_ref(), by_arity, |c, a| {
-            self.call_sub_value(c.clone(), a, true)
-        })
-    }
-
-    /// Engine-agnostic `minmax` over a list: collect candidates, fold the min and
-    /// the max (using the `:by` block — the genuine Category-B fork — through
-    /// `call`), and build the inclusive `min..max` range. Shared by the
-    /// interpreter (`builtin_minmax`) and the VM (`try_native_minmax`) so the
-    /// candidate / fold / range-build logic exists once. `by_arity` is the
-    /// resolved arity of `by` (1 = sort-key mapper, otherwise a comparator).
-    pub(crate) fn minmax_from_values_generic<F>(
-        positional: &[Value],
-        by: Option<&Value>,
-        by_arity: Option<usize>,
-        mut call: F,
-    ) -> Result<Value, RuntimeError>
-    where
-        F: FnMut(&Value, Vec<Value>) -> Result<Value, RuntimeError>,
-    {
-        let mut candidates = Vec::new();
-        if by.is_some() {
-            // When a comparator is provided, enumerate ranges fully
-            for arg in positional {
-                match arg {
-                    Value::Range(..)
-                    | Value::RangeExcl(..)
-                    | Value::RangeExclStart(..)
-                    | Value::RangeExclBoth(..)
-                    | Value::GenericRange { .. } => {
-                        let items = crate::runtime::utils::value_to_list(arg);
-                        candidates.extend(items);
-                    }
-                    _ if arg.as_list_items().is_some() => {
-                        candidates.extend(arg.as_list_items().unwrap().iter().cloned());
-                    }
-                    other => candidates.push(other.clone()),
-                }
-            }
-        } else {
-            for arg in positional {
-                Self::collect_minmax_candidates(arg, &mut candidates);
-            }
-        }
-        if candidates.is_empty() {
-            // Return Inf..-Inf for empty list (Raku behavior)
-            return Ok(Self::make_inclusive_range_value(
-                Value::Num(f64::INFINITY),
-                Value::Num(f64::NEG_INFINITY),
-            ));
-        }
-
-        if let Some(by_block) = by {
-            let by_arity = by_arity.unwrap_or(2);
-            let mut min_value = candidates[0].clone();
-            let mut max_value = candidates[0].clone();
-            for value in candidates.iter().skip(1) {
-                let cmp_min = if by_arity == 1 {
-                    let key_a = call(by_block, vec![value.clone()])?;
-                    let key_b = call(by_block, vec![min_value.clone()])?;
-                    crate::runtime::compare_values(&key_a, &key_b)
-                } else {
-                    let result = call(by_block, vec![value.clone(), min_value.clone()])?;
-                    Self::order_to_int(&result)
-                };
-                let cmp_max = if by_arity == 1 {
-                    let key_a = call(by_block, vec![value.clone()])?;
-                    let key_b = call(by_block, vec![max_value.clone()])?;
-                    crate::runtime::compare_values(&key_a, &key_b)
-                } else {
-                    let result = call(by_block, vec![value.clone(), max_value.clone()])?;
-                    Self::order_to_int(&result)
-                };
-                if cmp_min < 0 {
-                    min_value = value.clone();
-                }
-                if cmp_max > 0 {
-                    max_value = value.clone();
-                }
-            }
-            Ok(Self::make_inclusive_range_value(min_value, max_value))
-        } else {
-            let mut min_value = candidates[0].clone();
-            let mut max_value = candidates[0].clone();
-            for value in candidates.iter().skip(1) {
-                if crate::runtime::compare_values(value, &min_value) < 0 {
-                    min_value = value.clone();
-                }
-                if crate::runtime::compare_values(value, &max_value) > 0 {
-                    max_value = value.clone();
-                }
-            }
-            Ok(Self::make_inclusive_range_value(min_value, max_value))
-        }
-    }
-
-    /// Convert an Order/Int/Num value from a comparator callback to an integer.
-    fn order_to_int(result: &Value) -> i32 {
-        match result {
-            Value::Int(n) => (*n as i32).signum(),
-            Value::Num(n) => {
-                if *n < 0.0 {
-                    -1
-                } else if *n > 0.0 {
-                    1
-                } else {
-                    0
-                }
-            }
-            Value::Enum {
-                enum_type, value, ..
-            } if enum_type.resolve() == "Order" => match value {
-                crate::value::EnumValue::Int(n) => (*n as i32).signum(),
-                _ => 0,
-            },
-            _ => result.to_f64() as i32,
-        }
-    }
-
-    pub(super) fn builtin_shift(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            let msg = "X::TypeCheck::Argument: Calling shift(Any) will never work with declared signature ($)".to_string();
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("message".to_string(), Value::str(msg.clone()));
-            let ex = Value::make_instance(Symbol::intern("X::TypeCheck::Argument"), attrs);
-            let mut err = RuntimeError::new(msg);
-            err.exception = Some(Box::new(ex));
-            return Err(err);
-        }
-        if args.len() > 1 {
-            return Err(RuntimeError::new(format!(
-                "Too many positionals passed; expected 1 argument but got {}",
-                args.len()
-            )));
-        }
-        Ok(match args.first().cloned() {
-            Some(Value::Array(mut items, ..)) => {
-                if items.is_empty() {
-                    make_empty_array_failure("shift")
-                } else {
-                    Arc::make_mut(&mut items).remove(0)
-                }
-            }
-            _ => make_empty_array_failure("shift"),
-        })
-    }
-
-    pub(super) fn builtin_pop(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            let msg =
-                "Calling pop() will never work with signature of the proto ($, *%)".to_string();
-            let mut attrs = StdHashMap::new();
-            attrs.insert("message".to_string(), Value::str(msg.clone()));
-            let ex = Value::make_instance(Symbol::intern("X::TypeCheck::Argument"), attrs);
-            let mut err = RuntimeError::new(msg);
-            err.exception = Some(Box::new(ex));
-            return Err(err);
-        }
-        if args.len() > 1 {
-            return Err(RuntimeError::new(format!(
-                "Too many positionals passed to pop; expected 1 argument but got {}",
-                args.len()
-            )));
-        }
-        match args.first() {
-            Some(Value::Array(_, kind)) if kind.is_lazy() => {
-                return Err(RuntimeError::cannot_lazy("pop"));
-            }
-            _ => {}
-        }
-        Ok(match args.first().cloned() {
-            Some(Value::Array(mut items, ..)) => {
-                let items_mut = Arc::make_mut(&mut items);
-                if items_mut.is_empty() {
-                    make_empty_array_failure("pop")
-                } else {
-                    items_mut.pop().unwrap_or(Value::Nil)
-                }
-            }
-            _ => make_empty_array_failure("pop"),
-        })
-    }
-
-    pub(super) fn builtin_join(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let sep = args
-            .first()
-            .map(|v| v.to_string_value())
-            .unwrap_or_default();
-        // `join(sep, *@rest)` joins the flattened rest. Force any lazy lists
-        // (join is eager over the flattened list) then run the single shared
-        // `join_flat` helper — the same one the pure `native_function("join", ..)`
-        // path uses. The previous inline copy here drifted from native (it had a
-        // separate Range/Seq/itemized walk).
-        let mut rest = Vec::with_capacity(args.len().saturating_sub(1));
-        // `args.get(1..)` (not `&args[1..]`) so bare `join()` / `join(sep)` with no
-        // list args doesn't panic on an empty slice.
-        for v in args.get(1..).unwrap_or(&[]) {
-            if let Value::LazyList(list) = v {
-                // A genuinely-lazy `@`-array stays opaque (rendered `...` by
-                // `flat_val`) rather than forcing its capped prefix — this is the
-                // `"@a[]"` interpolation path (`join " ", @a`). (L2)
-                if list.in_array_context() && list.is_genuinely_lazy() {
-                    rest.push(v.clone());
-                } else {
-                    rest.push(Value::array(self.force_lazy_list(list)?));
-                }
-            } else {
-                rest.push(v.clone());
-            }
-        }
-        Ok(Value::str(
-            crate::builtins::join_flat(&sep, &rest).unwrap_or_default(),
-        ))
-    }
-
-    pub(super) fn builtin_list(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // `list` on a single Seq is a no-op (returns the Seq as-is)
-        if args.len() == 1
-            && let Value::Seq(_) = &args[0]
-        {
-            return Ok(args[0].clone());
-        }
-        let mut result = Vec::new();
-        for arg in args {
-            result.extend(Self::value_to_list(arg));
-        }
-        Ok(Value::array(result))
-    }
-
-    pub(super) fn builtin_cache(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // `cache` eagerly evaluates and caches the values into a List
-        let mut result = Vec::new();
-        for arg in args {
-            result.extend(Self::value_to_list(arg));
-        }
-        Ok(Value::array(result))
-    }
-
-    pub(super) fn builtin_flat(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // `flat(a, b, c)` flattens the argument *list*, so wrap the args in a
-        // `List` and run the single shared `flat` helper (the same one the pure
-        // `native_function("flat", ..)` path uses). The previous `flat_into`
-        // copy unconditionally flattened nested `[...]`, which over-flattened
-        // `flat(1, [2, [3, 4]], (5, 6))` (raku keeps the inner `[3 4]`).
-        // Onearg rule: `flat($(1,2,3))` un-itemizes its sole argument and
-        // descends, but `flat(0, $(1,2,3))` keeps each itemized arg single.
-        let items: Vec<Value> = if args.len() == 1 {
-            vec![crate::builtins::deitemize_flat_operand(&args[0])]
-        } else {
-            args.to_vec()
-        };
-        let list = Value::Array(
-            std::sync::Arc::new(crate::value::ArrayData::new(items)),
-            crate::value::ArrayKind::List,
-        );
-        let mut result = Vec::new();
-        crate::builtins::flat_val(&list, &mut result, true);
-        Ok(Value::Seq(std::sync::Arc::new(result)))
-    }
-
-    pub(super) fn builtin_slip(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let mut items = Vec::new();
-        for arg in args {
-            match arg {
-                Value::Array(elems, ..) => items.extend(elems.iter().cloned()),
-                Value::Seq(elems) => items.extend(elems.iter().cloned()),
-                Value::Slip(elems) => items.extend(elems.iter().cloned()),
-                Value::LazyList(ll) => {
-                    if ll.scan_spec.is_some() {
-                        items.extend(ll.force_scan_to(200_000));
-                    } else {
-                        let cached = ll.cache.lock().unwrap().clone().unwrap_or_default();
-                        items.extend(cached);
-                    }
-                }
-                other => items.push(other.clone()),
-            }
-        }
-        Ok(Value::slip(items))
-    }
-
-    pub(super) fn builtin_reverse(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // `reverse()` (empty) and `reverse(a, b, c)` reverse the argument *list*
-        // (so `reverse()` is `()`, not `Nil`). Only the single-arg form reverses a
-        // collection.
-        if args.len() != 1 {
-            let mut items: Vec<Value> = args.to_vec();
-            items.reverse();
-            return Ok(Value::Seq(Arc::new(items)));
-        }
-        // LazyIoLines must be materialized by the interpreter (native bails).
-        if let Value::LazyIoLines { handle, .. } = &args[0] {
-            let mut lines = Vec::new();
-            while let Some(line) = self.read_line_from_handle_value(handle)? {
-                lines.push(Value::str(line));
-            }
-            lines.reverse();
-            return Ok(Value::Seq(Arc::new(lines)));
-        }
-        // Single arg: delegate to the single shared native `reverse` (Array / Seq
-        // / Slip / Range / 1-D shaped / Str), instead of a drifting second copy
-        // that lacked Range/Slip/shaped arms.
-        crate::builtins::native_function(Symbol::intern("reverse"), args).unwrap_or(Ok(Value::Nil))
-    }
-
-    pub(super) fn builtin_sort(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // Extract named args (:k, :by) and positional args
-        let mut return_k = false;
-        let mut by_callable: Option<Value> = None;
-        let mut callable: Option<Value> = None;
-        let mut positional: Vec<Value> = Vec::new();
-
-        for arg in args {
-            match arg {
-                Value::Pair(key, val) if key == "k" => {
-                    return_k = val.truthy();
-                }
-                Value::Pair(key, val) if key == "by" => {
-                    by_callable = Some(val.as_ref().clone());
-                }
-                _ => positional.push(arg.clone()),
-            }
-        }
-
-        // A lazy/infinite source cannot be sorted: throw X::Cannot::Lazy instead
-        // of materializing it (matches raku). The comparator (a Sub) is never
-        // lazy, so scanning all positionals is safe.
-        if positional.iter().any(Self::is_lazy_for_coerce) {
-            return Err(RuntimeError::cannot_lazy("sort"));
-        }
-
-        // sort(comparator, list, ...) or sort(list, ...) or sort(items...)
-        if positional.len() >= 2 {
-            let first = &positional[0];
-            if matches!(first, Value::Sub(_) | Value::Routine { .. }) {
-                callable = Some(first.clone());
-                positional.remove(0);
-            }
-        }
-
-        // Prefer :by over positional callable
-        if by_callable.is_some() {
-            callable = by_callable;
-        }
-
-        // Flatten positional args into items
-        let mut items: Vec<Value> = Vec::new();
-        for arg in &positional {
-            if crate::runtime::utils::is_shaped_array(arg) {
-                items.extend(crate::runtime::utils::shaped_array_leaves(arg));
-            } else {
-                match arg {
-                    Value::Array(elems, ..) => items.extend(elems.iter().cloned()),
-                    Value::Seq(elems) => items.extend(elems.iter().cloned()),
-                    Value::Range(..)
-                    | Value::RangeExcl(..)
-                    | Value::RangeExclStart(..)
-                    | Value::RangeExclBoth(..)
-                    | Value::GenericRange { .. } => {
-                        items.extend(Self::value_to_list(arg));
-                    }
-                    other => items.push(other.clone()),
-                }
-            }
-        }
-
-        // Build sort args for the shared sort entry point.
-        let mut sort_args: Vec<Value> = Vec::new();
-        if let Some(c) = callable {
-            sort_args.push(c);
-        }
-        if return_k {
-            sort_args.push(Value::Pair("k".to_string(), Box::new(Value::Bool(true))));
-        }
-
-        if items.is_empty() && positional.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let mut caller = crate::runtime::methods_collection_ops::sort::InterpCaller(self);
-        crate::runtime::methods_collection_ops::sort::sort_value_generic(
-            &mut caller,
-            Value::array(items),
-            &sort_args,
-        )
-    }
-
-    pub(super) fn builtin_unique(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let mut positional = Vec::new();
-        let mut method_args = Vec::new();
-        for arg in args {
-            match arg {
-                Value::Pair(key, _) if key == "as" || key == "with" => {
-                    method_args.push(arg.clone());
-                }
-                _ => positional.push(arg.clone()),
-            }
-        }
-
-        if positional.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let target = if positional.len() == 1 {
-            positional[0].clone()
-        } else {
-            Value::array(positional)
-        };
-        self.call_method_with_values(target, "unique", method_args)
-    }
-
-    pub(super) fn builtin_repeated(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let mut positional = Vec::new();
-        let mut method_args = Vec::new();
-        for arg in args {
-            match arg {
-                Value::Pair(key, _) if key == "as" || key == "with" => {
-                    method_args.push(arg.clone());
-                }
-                _ => positional.push(arg.clone()),
-            }
-        }
-
-        if positional.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let target = if positional.len() == 1 {
-            positional[0].clone()
-        } else {
-            Value::array(positional)
-        };
-        self.call_method_with_values(target, "repeated", method_args)
-    }
-
-    pub(super) fn builtin_squish(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let mut positional = Vec::new();
-        let mut method_args = Vec::new();
-        for arg in args {
-            match arg {
-                Value::Pair(key, _) if key == "as" || key == "with" => {
-                    method_args.push(arg.clone());
-                }
-                _ => positional.push(arg.clone()),
-            }
-        }
-
-        if positional.is_empty() {
-            return Ok(Value::array(Vec::new()));
-        }
-
-        let target = if positional.len() == 1 {
-            positional[0].clone()
-        } else {
-            Value::array(positional)
-        };
-        self.call_method_with_values(target, "squish", method_args)
-    }
-
-    pub(super) fn builtin_map(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let func = args.first().cloned();
-        let arg_sources = self.pending_call_arg_sources.clone().unwrap_or_default();
-        // Check if the source is a single named array variable for rw writeback
-        let source_var = if args.len() == 2 {
-            arg_sources
-                .get(1)
-                .and_then(|entry| entry.as_ref())
-                .filter(|name| name.starts_with('@'))
-                .cloned()
-        } else {
-            None
-        };
-        let mut list_items = Vec::new();
-        for (idx, arg) in args.iter().skip(1).enumerate() {
-            // Check if this argument came from a $ variable (itemized container).
-            // Scalar variables don't have a sigil prefix in arg_sources, while
-            // @array and %hash variables start with '@' and '%' respectively.
-            let is_itemized = arg_sources
-                .get(idx + 1)
-                .and_then(|entry| entry.as_ref())
-                .is_some_and(|name| {
-                    !name.starts_with('@') && !name.starts_with('%') && !name.starts_with('&')
-                });
-            if !is_itemized && crate::runtime::utils::is_shaped_array(arg) {
-                list_items.extend(crate::runtime::utils::shaped_array_leaves(arg));
-            } else if is_itemized {
-                // Itemized containers (from $ variables) are NOT flattened
-                list_items.push(arg.clone());
-            } else {
-                match arg {
-                    Value::Array(items, ..) => list_items.extend(items.iter().cloned()),
-                    Value::Seq(items) => list_items.extend(items.iter().cloned()),
-                    Value::Hash(map) => {
-                        // Hashes in list context flatten to key-value Pairs
-                        for (k, v) in map.iter() {
-                            list_items.push(Value::Pair(k.clone(), Box::new(v.clone())));
-                        }
-                    }
-                    v if v.is_range() => {
-                        // Route ranges through the unified pull iterator so an
-                        // open-ended range (`1..Inf` == `Range(1, i64::MAX)`) is
-                        // truncated at the cap instead of panicking with
-                        // `capacity overflow` (ANALYSIS §8.2).
-                        list_items.extend(crate::runtime::value_iterator::materialize_capped(
-                            v,
-                            crate::runtime::utils::MAX_RANGE_EXPAND as usize,
-                        ));
-                    }
-                    other => list_items.push(other.clone()),
-                }
-            }
-        }
-        if let Some(var_name) = source_var {
-            let result = self.eval_map_over_items_rw(func, &mut list_items)?;
-            self.env.insert(var_name, Value::real_array(list_items));
-            Ok(result)
-        } else {
-            self.eval_map_over_items(func, list_items)
-        }
-    }
-
-    pub(super) fn builtin_grep(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // Parse named adverbs (:k, :v, :kv, :p) from args
-        let mut has_k = false;
-        let mut has_kv = false;
-        let mut has_p = false;
-        let mut positional_args: Vec<Value> = Vec::new();
-        for arg in args {
-            match arg {
-                Value::Pair(key, value) if key == "k" => has_k = value.truthy(),
-                Value::Pair(key, value) if key == "kv" => has_kv = value.truthy(),
-                Value::Pair(key, value) if key == "p" => has_p = value.truthy(),
-                Value::Pair(key, value) if key == "v" => {
-                    if !value.truthy() {
-                        return Err(RuntimeError::new(
-                            "X::Adverb: Unexpected adverb 'v' passed to grep",
-                        ));
-                    }
-                }
-                _ => positional_args.push(arg.clone()),
-            }
-        }
-        let args = &positional_args;
-        let func = args.first().cloned();
-        // Check if the matcher is a Bool — this is always an error.
-        // `grep $_ == 1, 1,2,3` passes a Bool as the matcher.
-        if let Some(Value::Bool(_)) = func.as_ref() {
-            return Err(RuntimeError::typed_msg(
-                "X::Match::Bool",
-                "Cannot use Bool as Matcher with '.match'. Did you mean to use $_ ~~ ... instead?",
-            ));
-        }
-        if args.len() == 1 && matches!(args[0], Value::Int(_) | Value::Num(_) | Value::Str(_)) {
-            return Err(RuntimeError::typed_msg(
-                "X::Match::Bool",
-                "Cannot use Bool as Matcher with '.match'. Did you mean to use $_ ~~ ... instead?",
-            ));
-        }
-        if args.len() == 2
-            && let Value::GenericRange { end, .. } = &args[1]
-        {
-            let end_f = end.to_f64();
-            if end_f.is_infinite() && end_f.is_sign_positive() {
-                let mut method_args: Vec<Value> = func.into_iter().collect();
-                if has_k {
-                    method_args.push(Value::Pair("k".to_string(), Box::new(Value::Bool(true))));
-                } else if has_kv {
-                    method_args.push(Value::Pair("kv".to_string(), Box::new(Value::Bool(true))));
-                } else if has_p {
-                    method_args.push(Value::Pair("p".to_string(), Box::new(Value::Bool(true))));
-                }
-                return self.call_method_with_values(args[1].clone(), "grep", method_args);
-            }
-        }
-        let mut list_items = Vec::new();
-        for arg in args.iter().skip(1) {
-            match arg {
-                // Only flatten List-kind arrays (from @-sigiled variables).
-                // Array-kind (from [...] literals) are kept as individual items.
-                Value::Array(items, kind)
-                    if matches!(
-                        kind,
-                        crate::value::ArrayKind::List | crate::value::ArrayKind::ItemList
-                    ) || (matches!(kind, crate::value::ArrayKind::Array) && args.len() == 2) =>
-                {
-                    list_items.extend(items.iter().cloned());
-                }
-                // A Seq (e.g. the result of `@a.sort`) is an iterable sequence and
-                // is flattened into grep's list, mirroring `map` (which already
-                // handles `Value::Seq`). Without this, `grep { ... }, @a.sort`
-                // greps over a single one-element list.
-                Value::Seq(items) => {
-                    list_items.extend(items.iter().cloned());
-                }
-                v if v.is_range() => {
-                    // Route ranges through the unified pull iterator so an
-                    // open-ended range (`1..Inf` == `Range(1, i64::MAX)`) is
-                    // truncated at the cap instead of panicking with
-                    // `capacity overflow` (ANALYSIS §8.2).
-                    list_items.extend(crate::runtime::value_iterator::materialize_capped(
-                        v,
-                        crate::runtime::utils::MAX_RANGE_EXPAND as usize,
-                    ));
-                }
-                other => list_items.push(other.clone()),
-            }
-        }
-        if has_k || has_kv || has_p {
-            let original_items = list_items.clone();
-            let filtered = self.eval_grep_over_items(func, list_items)?;
-            let indices = crate::runtime::methods_collection_ops::compute_grep_indices(
-                &original_items,
-                &filtered,
-            );
-            if has_k {
-                let idx_vals: Vec<Value> = indices.iter().map(|&i| Value::Int(i as i64)).collect();
-                Ok(Value::array(idx_vals))
-            } else if has_kv {
-                let items = if let Value::Array(items, ..) = &filtered {
-                    items.to_vec()
-                } else {
-                    vec![filtered]
-                };
-                let mut result = Vec::new();
-                for (i, item) in indices.iter().zip(items.iter()) {
-                    result.push(Value::Int(*i as i64));
-                    result.push(item.clone());
-                }
-                Ok(Value::array(result))
-            } else {
-                // :p
-                let items = if let Value::Array(items, ..) = &filtered {
-                    items.to_vec()
-                } else {
-                    vec![filtered]
-                };
-                let mut result = Vec::new();
-                for (i, item) in indices.iter().zip(items.iter()) {
-                    // The key is the Int index (`3 => v`), not a Str ("3" => v).
-                    result.push(Value::ValuePair(
-                        Box::new(Value::Int(*i as i64)),
-                        Box::new(item.clone()),
-                    ));
-                }
-                Ok(Value::array(result))
-            }
-        } else {
-            self.eval_grep_over_items(func, list_items)
-        }
-    }
-
-    /// `snip(matcher, +values)` — split a list at positions where the matcher stops matching.
-    pub(super) fn builtin_snip(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Err(RuntimeError::new("Too few positionals passed to 'snip'"));
-        }
-        let matcher = args[0].clone();
-        let mut list_items = Vec::new();
-        for arg in args.iter().skip(1) {
-            match arg {
-                Value::Array(items, ..) => list_items.extend(items.iter().cloned()),
-                other => list_items.push(other.clone()),
-            }
-        }
-        self.eval_snip(matcher, list_items)
-    }
-
-    /// Core snip implementation shared by both sub and method forms.
-    pub(super) fn eval_snip(
-        &mut self,
-        matcher: Value,
-        items: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        // Extract the list of matchers: if matcher is a list/array of callables/types,
-        // use them in sequence; otherwise treat as a single matcher.
-        let matchers: Vec<Value> = match &matcher {
-            Value::Array(elems, ..) => elems.iter().cloned().collect(),
-            Value::Seq(elems) | Value::Slip(elems) => elems.iter().cloned().collect(),
-            other => vec![other.clone()],
-        };
-
-        let mut result_groups: Vec<Value> = Vec::new();
-        let mut current_group: Vec<Value> = Vec::new();
-        let mut matcher_idx: usize = 0;
-
-        for item in items {
-            let current_matcher = matchers.get(matcher_idx);
-            let matched = if let Some(m) = current_matcher {
-                self.snip_matches(&item, m)?
-            } else {
-                // No more matchers — everything goes into the last group
-                true
-            };
-
-            if matched {
-                current_group.push(item);
-            } else {
-                // Snip here: save current group and start a new one
-                result_groups.push(Value::array(current_group));
-                current_group = vec![item];
-                matcher_idx += 1;
-            }
-        }
-        // Push the final group
-        if !current_group.is_empty() {
-            result_groups.push(Value::array(current_group));
-        }
-
-        Ok(Value::array(result_groups))
-    }
-
-    /// Check if a value matches a snip matcher (Callable or type object).
-    fn snip_matches(&mut self, item: &Value, matcher: &Value) -> Result<bool, RuntimeError> {
-        if matches!(matcher, Value::Sub(_)) {
-            Ok(self
-                .call_sub_value(matcher.clone(), vec![item.clone()], true)?
-                .truthy())
-        } else {
-            Ok(self.smart_match(item, matcher))
-        }
-    }
-
-    /// If `v` is a Buf/Blob, return its element bytes as a list of Int values.
-    /// Raku iterates a Blob as its bytes, so list methods (`map`/`grep`/`first`)
-    /// must expand it rather than treating the whole Blob as a single element.
-    /// (Note: list-*assignment* keeps a Blob as one element — `my @a = $buf` —
-    /// so this is used only by the iterating methods, not `value_to_list`.)
-    pub(crate) fn buf_as_byte_items(v: &Value) -> Option<Vec<Value>> {
-        if let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = v
-        {
-            let cn = class_name.resolve();
-            let is_blob = cn == "Buf"
-                || cn == "Blob"
-                || cn == "utf8"
-                || cn == "utf16"
-                || cn == "utf32"
-                || cn.starts_with("Buf[")
-                || cn.starts_with("Blob[")
-                || cn.starts_with("buf")
-                || cn.starts_with("blob");
-            if is_blob && let Some(Value::Array(bytes, ..)) = attributes.as_map().get("bytes") {
-                return Some(bytes.iter().cloned().collect());
-            }
-        }
-        None
-    }
-
-    pub(super) fn builtin_first(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        // Separate named args (Pairs) from positional args
-        let mut positional = Vec::new();
-        let mut has_v = false;
-        let mut has_neg_v = false;
-        let mut has_end = false;
-        let mut has_k = false;
-        let mut has_kv = false;
-        let mut has_p = false;
-        for arg in args {
-            match arg {
-                Value::Pair(key, value) if key == "v" => {
-                    if value.truthy() {
-                        has_v = true;
-                    } else {
-                        has_neg_v = true;
-                    }
-                }
-                Value::Pair(key, value) if key == "end" => {
-                    if value.truthy() {
-                        has_end = true;
-                    }
-                }
-                Value::Pair(key, value) if key == "k" => {
-                    has_k = value.truthy();
-                }
-                Value::Pair(key, value) if key == "kv" => {
-                    has_kv = value.truthy();
-                }
-                Value::Pair(key, value) if key == "p" => {
-                    has_p = value.truthy();
-                }
-                _ => positional.push(arg.clone()),
-            }
-        }
-        if has_neg_v {
-            return Err(RuntimeError::new(
-                "Throwing `:!v` on first is not supported",
-            ));
-        }
-        let _ = has_v; // :v is the default behavior
-        // Check for conflicting adverbs (X::Adverb)
-        {
-            let adverb_count = has_k as u8 + has_kv as u8 + has_p as u8;
-            if adverb_count > 1 {
-                let mut names = Vec::new();
-                if has_k {
-                    names.push("k");
-                }
-                if has_kv {
-                    names.push("kv");
-                }
-                if has_p {
-                    names.push("p");
-                }
-                let adverb_list = names
-                    .iter()
-                    .map(|n| format!("'{}'", n))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                let mut err = RuntimeError::new(format!(
-                    "Unsupported combination of adverbs ({}) passed to first on\n'List'.",
-                    adverb_list
-                ));
-                err.exception = Some(Box::new(Value::make_instance(
-                    crate::symbol::Symbol::intern("X::Adverb"),
-                    std::collections::HashMap::new(),
-                )));
-                return Err(err);
-            }
-        }
-        // Check for Bool matcher (X::Match::Bool)
-        if matches!(positional.first(), Some(Value::Bool(_))) {
-            let mut err = RuntimeError::new("Cannot use Bool as a matcher");
-            err.exception = Some(Box::new(Value::make_instance(
-                Symbol::intern("X::Match::Bool"),
-                std::collections::HashMap::new(),
-            )));
-            return Err(err);
-        }
-        let func = positional.first().cloned();
-        let mut list_items = Vec::new();
-        for arg in positional.iter().skip(1) {
-            match arg {
-                Value::Array(items, ..) => list_items.extend(items.iter().cloned()),
-                v if v.is_range() => {
-                    // Route ranges through the unified pull iterator so an
-                    // open-ended range (`1..Inf` == `Range(1, i64::MAX)`) is
-                    // truncated at the cap instead of panicking with
-                    // `capacity overflow` (ANALYSIS §8.2).
-                    list_items.extend(crate::runtime::value_iterator::materialize_capped(
-                        v,
-                        crate::runtime::utils::MAX_RANGE_EXPAND as usize,
-                    ));
-                }
-                other => list_items.push(other.clone()),
-            }
-        }
-        if let Some((idx, value)) = self.find_first_match_over_items(func, &list_items, has_end)? {
-            return Ok(format_first_result(idx, value, has_k, has_kv, has_p));
-        }
-        Ok(Value::Nil)
-    }
-
-    /// Find a variable name in the current environment whose value is identical
-    /// (by Arc pointer) to the given value. Used for :into writeback.
-    fn find_var_by_identity(&self, value: &Value) -> Option<String> {
-        match value {
-            Value::Hash(target_arc) => {
-                for (name, env_val) in self.env().iter() {
-                    if let Value::Hash(env_arc) = env_val
-                        && std::sync::Arc::ptr_eq(target_arc, env_arc)
-                    {
-                        return Some(name.resolve());
-                    }
-                }
-                None
-            }
-            Value::Bag(target_arc, _) => {
-                for (name, env_val) in self.env().iter() {
-                    if let Value::Bag(env_arc, _) = env_val
-                        && std::sync::Arc::ptr_eq(target_arc, env_arc)
-                    {
-                        return Some(name.resolve());
-                    }
-                }
-                None
-            }
-            Value::Mix(target_arc, _) => {
-                for (name, env_val) in self.env().iter() {
-                    if let Value::Mix(env_arc, _) = env_val
-                        && std::sync::Arc::ptr_eq(target_arc, env_arc)
-                    {
-                        return Some(name.resolve());
-                    }
-                }
-                None
-            }
-            _ => None,
-        }
-    }
-
-    pub(super) fn builtin_classify(
-        &mut self,
-        name: &str,
-        args: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        fn as_items(value: &Value) -> Option<Vec<Value>> {
-            match value {
-                Value::Array(items, ..) => Some(items.iter().cloned().collect()),
-                Value::Seq(items) | Value::Slip(items) => Some(items.iter().cloned().collect()),
-                _ => None,
-            }
-        }
-
-        fn mapper_path(value: Value) -> Vec<Value> {
-            as_items(&value).unwrap_or_else(|| vec![value])
-        }
-
-        /// Returns (paths, is_multi_level).
-        /// `is_multi_level` is true when the mapper returned a list whose
-        /// elements are themselves lists (multi-level paths), as opposed to
-        /// a flat list of single category names.
-        fn mapper_categories(value: Value) -> (Vec<Vec<Value>>, bool) {
-            if let Some(items) = as_items(&value) {
-                if items.is_empty() {
-                    return (Vec::new(), false);
-                }
-                let has_list_element = items.iter().any(|i| as_items(i).is_some());
-                let mut out = Vec::new();
-                for item in items {
-                    out.push(mapper_path(item));
-                }
-                (out, has_list_element)
-            } else {
-                (vec![vec![value]], false)
-            }
-        }
-
-        fn mixed_level_error(name: &str) -> RuntimeError {
-            make_mixed_level_err(name)
-        }
-
-        fn make_mixed_level_err(name: &str) -> RuntimeError {
-            let kind = if name == "categorize" {
-                "categorization"
-            } else {
-                "classification"
-            };
-            let msg = format!(
-                "mapper on {}-list computed to an item with different number of elements in it than previous items, which cannot be used because all values need to have the same number of elements. Mixed-level {} is not supported.",
-                name, kind
-            );
-            let mut attrs = HashMap::new();
-            attrs.insert("message".to_string(), Value::str(msg.clone()));
-            let exception =
-                Value::make_instance(Symbol::intern("X::Invalid::ComputedValue"), attrs);
-            let mut err = RuntimeError::new(msg);
-            err.exception = Some(Box::new(exception));
-            err
-        }
-
-        fn insert_nested_bucket(
-            buckets: &mut HashMap<String, Value>,
-            path: &[Value],
-            item: Value,
-            name: &str,
-        ) -> Result<(), RuntimeError> {
-            if path.is_empty() {
-                return Ok(());
-            }
-            let key = path[0].to_string_value();
-            if path.len() == 1 {
-                let entry = buckets
-                    .entry(key)
-                    .or_insert_with(|| Value::real_array(Vec::new()));
-                match entry {
-                    Value::Array(values, ..) => {
-                        Arc::make_mut(values).push(item);
-                        Ok(())
-                    }
-                    Value::Hash(_) => Err(mixed_level_error(name)),
-                    _ => {
-                        *entry = Value::real_array(vec![item]);
-                        Ok(())
-                    }
-                }
-            } else {
-                let entry = buckets
-                    .entry(key)
-                    .or_insert_with(|| Value::hash(HashMap::new()));
-                match entry {
-                    Value::Hash(map) => {
-                        let map = Arc::make_mut(map);
-                        insert_nested_bucket(map, &path[1..], item, name)
-                    }
-                    Value::Array(..) => Err(mixed_level_error(name)),
-                    _ => {
-                        *entry = Value::hash(HashMap::new());
-                        if let Value::Hash(map) = entry {
-                            let map = Arc::make_mut(map);
-                            insert_nested_bucket(map, &path[1..], item, name)
-                        } else {
-                            Ok(())
-                        }
-                    }
-                }
-            }
-        }
-
-        // Extract variable names from arg_sources for :into writeback
-        let arg_sources = self.take_pending_call_arg_sources();
-        let mut mapper: Option<Value> = None;
-        let mut as_mapper: Option<Value> = None;
-        let mut into_target_raw: Option<Value> = None;
-        let mut into_target: Option<Value> = None;
-        let mut into_varname: Option<String> = None;
-        let mut positional: Vec<Value> = Vec::new();
-        for (i, arg) in args.iter().enumerate() {
-            let fetched_arg = self.auto_fetch_proxy(arg)?;
-            match &fetched_arg {
-                Value::Pair(key, value) if key == "as" => {
-                    as_mapper = Some(self.auto_fetch_proxy(value)?)
-                }
-                Value::Pair(key, value) if key == "into" => {
-                    // Look up the variable name from arg_sources metadata
-                    // The compiler encodes FatArrow args as "key=varname"
-                    if let Some(ref sources) = arg_sources
-                        && let Some(Some(source)) = sources.get(i)
-                        && let Some(varname) = source.strip_prefix("into=")
-                    {
-                        into_varname = Some(varname.to_string());
-                    }
-                    // Fallback: search env for a variable with matching identity
-                    if into_varname.is_none() {
-                        into_varname = self.find_var_by_identity(value);
-                    }
-                    into_target_raw = Some(*value.clone());
-                    into_target = Some(self.auto_fetch_proxy(value)?);
-                }
-                _ => {
-                    if mapper.is_none() {
-                        mapper = Some(fetched_arg);
-                    } else {
-                        positional.push(fetched_arg);
-                    }
-                }
-            }
-        }
-        let Some(mapper) = mapper else {
-            return Ok(into_target.unwrap_or_else(|| Value::hash(HashMap::new())));
-        };
-
-        // Check for lazy lists — classify/categorize cannot operate on lazy lists
-        for arg in &positional {
-            let is_lazy = matches!(arg, Value::LazyList(_));
-            if is_lazy {
-                let mut attrs = HashMap::new();
-                attrs.insert(
-                    "action".to_string(),
-                    Value::str_from(if name == "categorize" {
-                        ".categorize-list"
-                    } else {
-                        ".classify-list"
-                    }),
-                );
-                attrs.insert(
-                    "message".to_string(),
-                    Value::str(format!("Cannot {}-list a lazy list", name)),
-                );
-                let exception = Value::make_instance(Symbol::intern("X::Cannot::Lazy"), attrs);
-                let mut err = RuntimeError::new("X::Cannot::Lazy");
-                err.exception = Some(Box::new(exception));
-                return Err(err);
-            }
-        }
-
-        // Flatten list/array arguments into individual items and force lazy inputs.
-        let mut items = Vec::new();
-        for arg in &positional {
-            match arg {
-                Value::Array(values, ..) => items.extend(values.iter().cloned()),
-                Value::Seq(values) | Value::Slip(values) => items.extend(values.iter().cloned()),
-                Value::LazyList(ll) => items.extend(self.force_lazy_list_bridge(ll)?),
-                v if v.is_range() => items.extend(crate::runtime::utils::value_to_list(v)),
-                other => items.push(other.clone()),
-            }
-        }
-
-        let mut buckets: HashMap<String, Value> = match into_target.as_ref() {
-            Some(Value::Hash(map)) => map.as_ref().map.clone(),
-            _ => HashMap::new(),
-        };
-        let mut bag_counts: Option<HashMap<String, i64>> = match into_target.as_ref() {
-            Some(Value::Bag(b, _)) => Some(crate::runtime::utils::bag_counts_as_i64(&b.counts)),
-            _ => None,
-        };
-        let mut mix_counts: Option<HashMap<String, f64>> = match into_target.as_ref() {
-            Some(Value::Mix(m, _)) => Some(m.weights.clone()),
-            _ => None,
-        };
-
-        // Track classification level depth across all items for mixed-level detection
-        let mut expected_level: Option<usize> = None;
-        let mut expected_multi_level: Option<bool> = None;
-
-        for item in &items {
-            let mapped = match &mapper {
-                Value::Sub(_) | Value::WeakSub(_) | Value::Routine { .. } => {
-                    self.call_sub_value(mapper.clone(), vec![item.clone()], true)?
-                }
-                Value::Hash(map) => map
-                    .get(&item.to_string_value())
-                    .cloned()
-                    .unwrap_or(Value::Nil),
-                Value::Array(values, ..) => {
-                    let idx = crate::runtime::to_int(item);
-                    if idx < 0 {
-                        Value::Nil
-                    } else {
-                        values.get(idx as usize).cloned().unwrap_or(Value::Nil)
-                    }
-                }
-                _ => Value::Nil,
-            };
-            let mapped = match mapped {
-                Value::LazyList(ll) => Value::array(self.force_lazy_list_bridge(&ll)?),
-                other => other,
-            };
-
-            let mapped_item = if let Some(as_fn) = &as_mapper {
-                self.call_sub_value(as_fn.clone(), vec![item.clone()], true)?
-            } else {
-                item.clone()
-            };
-
-            if bag_counts.is_some() || mix_counts.is_some() {
-                let paths = if name == "categorize" {
-                    let (p, _) = mapper_categories(mapped);
-                    p
-                } else {
-                    let path = mapper_path(mapped);
-                    if path.is_empty() {
-                        Vec::new()
-                    } else {
-                        vec![path]
-                    }
-                };
-                for path in paths {
-                    if path.len() != 1 {
-                        return Err(RuntimeError::new(format!(
-                            "X::Invalid::ComputedValue: multi-level {}",
-                            if name == "categorize" {
-                                "categorization"
-                            } else {
-                                "classification"
-                            }
-                        )));
-                    }
-                    let key = path[0].to_string_value();
-                    if let Some(counts) = bag_counts.as_mut() {
-                        *counts.entry(key.clone()).or_insert(0) += 1;
-                    }
-                    if let Some(counts) = mix_counts.as_mut() {
-                        *counts.entry(key).or_insert(0.0) += 1.0;
-                    }
-                }
-                continue;
-            }
-
-            let (paths, is_multi_level) = if name == "categorize" {
-                mapper_categories(mapped)
-            } else {
-                let path = mapper_path(mapped);
-                let ml = path.len() > 1;
-                let v = if path.is_empty() {
-                    Vec::new()
-                } else {
-                    vec![path]
-                };
-                (v, ml)
-            };
-            if !paths.is_empty() {
-                if let Some(prev_ml) = expected_multi_level {
-                    if is_multi_level != prev_ml {
-                        return Err(make_mixed_level_err(name));
-                    }
-                } else {
-                    expected_multi_level = Some(is_multi_level);
-                }
-            }
-            for path in &paths {
-                if let Some(expected) = expected_level {
-                    if path.len() != expected {
-                        return Err(make_mixed_level_err(name));
-                    }
-                } else {
-                    expected_level = Some(path.len());
-                }
-                insert_nested_bucket(&mut buckets, path, mapped_item.clone(), name)?;
-            }
-        }
-
-        // Write back result to the caller's variable if :into was specified
-        {
-            let has_proxy = into_target_raw
-                .as_ref()
-                .is_some_and(|value| matches!(value, Value::Proxy { .. }));
-            let has_varname = into_varname.is_some();
-            if has_proxy || has_varname {
-                let mut updated: Option<Value> = None;
-                if let Some(counts) = bag_counts.clone() {
-                    updated = Some(Value::bag(counts));
-                } else if let Some(counts) = mix_counts.clone() {
-                    updated = Some(Value::mix(counts));
-                } else if into_target.is_some() {
-                    updated = Some(Value::hash(buckets.clone()));
-                }
-                if let Some(new_value) = updated {
-                    if has_proxy {
-                        let _ = self.assign_proxy_lvalue(into_target_raw.unwrap(), new_value)?;
-                    } else if let Some(ref vname) = into_varname {
-                        self.env_mut().insert(vname.clone(), new_value);
-                    }
-                }
-            }
-        }
-
-        if let Some(counts) = bag_counts {
-            return Ok(Value::bag(counts));
-        }
-        if let Some(counts) = mix_counts {
-            return Ok(Value::mix(counts));
-        }
-
-        Ok(Value::hash(buckets))
-    }
-
-    /// `cross(@a, @b, ...)` — Cartesian product of lists.
-    /// With `with => &op`, applies the operator to each pair instead of making tuples.
-    pub(super) fn builtin_cross(&mut self, args: Vec<Value>) -> Result<Value, RuntimeError> {
-        let mut lists: Vec<Vec<Value>> = Vec::new();
-        let mut with_func: Option<Value> = None;
-
-        for arg in &args {
-            match arg {
-                Value::Pair(k, v) if k.as_str() == "with" => {
-                    with_func = Some(v.as_ref().clone());
-                }
-                _ => {
-                    let mut values = super::utils::value_to_list(arg);
-                    if values.len() == 1
-                        && let Some(single) = values.first()
-                    {
-                        match single {
-                            Value::Array(items, _) => {
-                                values = items.as_ref().clone().items;
-                            }
-                            Value::Seq(items) | Value::Slip(items) => {
-                                values = items.as_ref().clone();
-                            }
-                            _ => {}
-                        }
-                    }
-                    lists.push(values);
-                }
-            }
-        }
-
-        if lists.is_empty() {
-            return Ok(Value::Seq(std::sync::Arc::new(vec![])));
-        }
-
-        // Compute Cartesian product iteratively
-        let mut result: Vec<Vec<Value>> = vec![vec![]];
-        for list in &lists {
-            let mut new_result = Vec::new();
-            for combo in &result {
-                for item in list {
-                    let mut new_combo = combo.clone();
-                    new_combo.push(item.clone());
-                    new_result.push(new_combo);
-                }
-            }
-            result = new_result;
-        }
-
-        // Apply `with` function or create tuples. `cross` returns a Seq (so
-        // `.^name` is Seq, `.raku` shows `.Seq`), matching Rakudo and the `X`
-        // metaop.
-        if let Some(func) = with_func {
-            let mut final_result = Vec::new();
-            for combo in result {
-                let val = self.call_sub_value(func.clone(), combo, false)?;
-                final_result.push(val);
-            }
-            Ok(Value::Seq(std::sync::Arc::new(final_result)))
-        } else {
-            // Return as list of lists (tuples)
-            let tuples: Vec<Value> = result.into_iter().map(Value::array).collect();
-            Ok(Value::Seq(std::sync::Arc::new(tuples)))
-        }
-    }
-
-    pub(super) fn builtin_roundrobin(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Ok(Value::Seq(std::sync::Arc::new(Vec::new())));
-        }
-
-        // Implement Raku's single-arg rule (+@lol): when called with a single
-        // iterable arg, iterate it to get the list of streams.
-        let effective_args: Vec<Value> = if args.len() == 1 {
-            match &args[0] {
-                Value::Array(items, kind) if kind.is_itemized() => args.to_vec(),
-                Value::Array(items, _) => items.iter().cloned().collect(),
-                Value::Seq(items) | Value::Slip(items) => items.iter().cloned().collect(),
-                _ => args.to_vec(),
-            }
-        } else {
-            args.to_vec()
-        };
-
-        if effective_args.is_empty() {
-            return Ok(Value::Seq(std::sync::Arc::new(Vec::new())));
-        }
-
-        let streams: Vec<Vec<Value>> = effective_args
-            .iter()
-            .map(|arg| match arg {
-                Value::Capture { positional, named }
-                    if named.is_empty() && positional.len() == 1 =>
-                {
-                    vec![arg.clone()]
-                }
-                Value::Array(items, kind) if kind.is_itemized() => vec![arg.clone()],
-                Value::Array(items, _) => items.iter().cloned().collect(),
-                Value::Seq(items) | Value::Slip(items) => items.iter().cloned().collect(),
-                Value::Range(a, b) => (*a..=*b).map(Value::Int).collect(),
-                Value::RangeExcl(a, b) => (*a..*b).map(Value::Int).collect(),
-                v if v.is_range() => crate::runtime::utils::value_to_list(v),
-                other => vec![other.clone()],
-            })
-            .collect();
-
-        let mut indices = vec![0usize; streams.len()];
-        let mut rounds = Vec::new();
-        loop {
-            let mut tuple = Vec::new();
-            let mut progressed = false;
-            for (i, stream) in streams.iter().enumerate() {
-                if indices[i] < stream.len() {
-                    tuple.push(stream[indices[i]].clone());
-                    indices[i] += 1;
-                    progressed = true;
-                }
-            }
-            if !progressed {
-                break;
-            }
-            rounds.push(Value::array(tuple));
-        }
-
-        Ok(Value::Seq(std::sync::Arc::new(rounds)))
-    }
-
-    /// `duckmap(&block, \obj)` — apply block to each element; on type mismatch
-    /// descend recursively into iterables, or return the element unchanged.
-    pub(super) fn builtin_duckmap(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.len() < 2 {
-            return Err(RuntimeError::new("duckmap requires a block and an object"));
-        }
-        let block = args[0].clone();
-        let obj = args[1].clone();
-        self.duckmap_iterate(&block, &obj)
-    }
-
-    /// `deepmap(&block, \obj)` — apply block to every leaf element, preserving structure.
-    pub(super) fn builtin_deepmap(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.len() < 2 {
-            return Err(RuntimeError::new("deepmap requires a block and an object"));
-        }
-        let block = args[0].clone();
-        let obj = args[1].clone();
-        self.deepmap_iterate(&block, &obj)
-    }
-
-    /// Iterate over the elements of a value, applying duckmap to each.
-    /// This is the entry point for both the method and function forms.
-    pub(crate) fn duckmap_iterate(
-        &mut self,
-        block: &Value,
-        target: &Value,
-    ) -> Result<Value, RuntimeError> {
-        match target {
-            Value::Array(items, kind) => {
-                let mut result = Vec::new();
-                for item in items.iter() {
-                    match self.duckmap_element(block, item) {
-                        Ok(v) => result.push(v),
-                        Err(e) if e.is_next() => continue,
-                        Err(e) if e.is_last() => break,
-                        Err(e) => return Err(e),
-                    }
-                }
-                if kind.is_real_array() {
-                    Ok(Value::real_array(result))
-                } else {
-                    Ok(Value::array(result))
-                }
-            }
-            Value::Seq(items) => {
-                let mut result = Vec::new();
-                for item in items.iter() {
-                    match self.duckmap_element(block, item) {
-                        Ok(v) => result.push(v),
-                        Err(e) if e.is_next() => continue,
-                        Err(e) if e.is_last() => break,
-                        Err(e) => return Err(e),
-                    }
-                }
-                Ok(Value::Seq(std::sync::Arc::new(result)))
-            }
-            Value::Hash(map) => {
-                let mut result = std::collections::HashMap::new();
-                for (k, v) in map.iter() {
-                    match self.duckmap_element(block, v) {
-                        Ok(mapped) => {
-                            result.insert(k.clone(), mapped);
-                        }
-                        Err(e) if e.is_next() => continue,
-                        Err(e) if e.is_last() => break,
-                        Err(e) => return Err(e),
-                    }
-                }
-                Ok(Value::Hash(Value::hash_arc(result)))
-            }
-            // Single non-iterable value: try the block on it directly
-            _ => self.duckmap_element(block, target),
-        }
-    }
-
-    /// Recursively apply a block to every leaf element, preserving structure.
-    pub(crate) fn deepmap_iterate(
-        &mut self,
-        block: &Value,
-        target: &Value,
-    ) -> Result<Value, RuntimeError> {
-        self.deepmap_iterate_inner(block, target, false)
-    }
-
-    /// Call the deepmap block on a leaf element through a transient
-    /// `ContainerRef` cell, so a mutating callable (`++*`, `*--`, `$_++`)
-    /// writes through — Raku's `deepmap` passes each leaf as a *container*
-    /// and mutations are visible in the source structure. Returns the
-    /// block's (decontainerized) result plus the cell's post-call value for
-    /// the caller to write back into the source slot.
-    fn deepmap_leaf_call(
-        &mut self,
-        block: &Value,
-        leaf: &Value,
-    ) -> Result<(Value, Value), RuntimeError> {
-        let cell = std::sync::Arc::new(std::sync::Mutex::new(leaf.clone()));
-        let res = self.call_sub_value(
-            block.clone(),
-            vec![Value::ContainerRef(cell.clone())],
-            false,
-        )?;
-        let new_val = cell.lock().unwrap().clone();
-        Ok((res.deref_container(), new_val))
-    }
-
-    /// Inner recursive helper. `itemize_result` is true for nested calls
-    /// so that sublists get wrapped in Scalar containers.
-    fn deepmap_iterate_inner(
-        &mut self,
-        block: &Value,
-        target: &Value,
-        itemize_result: bool,
-    ) -> Result<Value, RuntimeError> {
-        match target {
-            // Type objects (e.g. Array, Hash) — return as-is to avoid hanging
-            Value::Package(_) => Ok(target.clone()),
-            Value::Array(items, kind) => {
-                // A sublist is itemized (wrapped in a Scalar container) only when
-                // its *parent* is a List, not when the parent is a real Array.
-                // Compare Rakudo: `(1,[2,3]).deepmap(*+1)` -> `(2, $[3, 4])` but
-                // `[1,[2,3]].deepmap(*+1)` -> `[2, [3, 4]]`.
-                let child_itemize = !kind.is_real_array();
-                let mut result = Vec::new();
-                for (idx, item) in items.iter().enumerate() {
-                    let is_leaf = !matches!(
-                        item,
-                        Value::Package(_) | Value::Array(..) | Value::Seq(_) | Value::Hash(_)
-                    );
-                    if is_leaf {
-                        match self.deepmap_leaf_call(block, item) {
-                            Ok((v, new_src)) => {
-                                if new_src != *item {
-                                    // Write the mutated leaf back into the source
-                                    // array in place so all holders of the Arc see
-                                    // it (Raku container semantics).
-                                    // SAFETY: aliased in-place mutation of a shared
-                                    // container; see `arc_contents_mut`.
-                                    unsafe {
-                                        crate::value::arc_contents_mut(items).items[idx] = new_src;
-                                    }
-                                }
-                                result.push(v);
-                            }
-                            Err(e) if e.is_next() => continue,
-                            Err(e) => return Err(e),
-                        }
-                        continue;
-                    }
-                    match self.deepmap_iterate_inner(block, item, child_itemize) {
-                        Ok(v) => result.push(v),
-                        Err(e) if e.is_next() => continue,
-                        Err(e) => return Err(e),
-                    }
-                }
-                let mut use_real_array = kind.is_real_array();
-                // If the source array has a type constraint (e.g. `my Str @a`)
-                // and the mapped values don't conform, downgrade to a List.
-                if use_real_array && let Some(info) = self.container_type_metadata(target) {
-                    let vt = &info.value_type;
-                    if !vt.is_empty() && result.iter().any(|v| !v.isa_check(vt)) {
-                        use_real_array = false;
-                    }
-                }
-                let arr_kind = if use_real_array {
-                    if itemize_result {
-                        crate::value::ArrayKind::ItemArray
-                    } else {
-                        crate::value::ArrayKind::Array
-                    }
-                } else if itemize_result {
-                    crate::value::ArrayKind::ItemList
-                } else {
-                    crate::value::ArrayKind::List
-                };
-                Ok(Value::Array(
-                    std::sync::Arc::new(crate::value::ArrayData::new(result)),
-                    arr_kind,
-                ))
-            }
-            Value::Seq(items) => {
-                let mut result = Vec::new();
-                for item in items.iter() {
-                    match self.deepmap_iterate_inner(block, item, true) {
-                        Ok(v) => result.push(v),
-                        Err(e) if e.is_next() => continue,
-                        Err(e) => return Err(e),
-                    }
-                }
-                if itemize_result {
-                    // Itemize the result as a list
-                    Ok(Value::Array(
-                        std::sync::Arc::new(crate::value::ArrayData::new(result)),
-                        crate::value::ArrayKind::ItemList,
-                    ))
-                } else {
-                    Ok(Value::Seq(std::sync::Arc::new(result)))
-                }
-            }
-            Value::Hash(map) => {
-                let mut result = std::collections::HashMap::new();
-                for (k, v) in map.iter() {
-                    let is_leaf = !matches!(
-                        v,
-                        Value::Package(_) | Value::Array(..) | Value::Seq(_) | Value::Hash(_)
-                    );
-                    if is_leaf {
-                        match self.deepmap_leaf_call(block, v) {
-                            Ok((Value::Slip(items), _)) if items.is_empty() => continue,
-                            Ok((val, new_src)) => {
-                                if new_src != *v {
-                                    // Write the mutated leaf back into the source
-                                    // hash in place (see the Array arm).
-                                    // SAFETY: aliased in-place mutation of a shared
-                                    // container; see `arc_contents_mut`.
-                                    unsafe {
-                                        crate::value::arc_contents_mut(map)
-                                            .map
-                                            .insert(k.clone(), new_src);
-                                    }
-                                }
-                                result.insert(k.clone(), val);
-                            }
-                            Err(e) if e.is_next() => continue,
-                            Err(e) => return Err(e),
-                        }
-                        continue;
-                    }
-                    match self.deepmap_iterate_inner(block, v, true) {
-                        Ok(Value::Slip(items)) if items.is_empty() => {
-                            // Empty slip means the block returned Empty;
-                            // drop the key from the result hash.
-                            continue;
-                        }
-                        Ok(val) => {
-                            result.insert(k.clone(), val);
-                        }
-                        Err(e) if e.is_next() => continue,
-                        Err(e) => return Err(e),
-                    }
-                }
-                Ok(Value::Hash(Value::hash_arc(result)))
-            }
-            // Leaf value: apply the block (through a transient container so
-            // mutating callables write through to a bare top-level leaf too).
-            _ => self.deepmap_leaf_call(block, target).map(|(v, _)| v),
-        }
-    }
-
-    /// `nodemap` — apply a block to each element without descending into sublists.
-    pub(crate) fn nodemap_iterate(
-        &mut self,
-        block: &Value,
-        target: &Value,
-    ) -> Result<Value, RuntimeError> {
-        match target {
-            // nodemap always returns a List, even from a real Array or a Seq.
-            // Compare Rakudo: `[2,3].nodemap(*+1).WHAT` is `List`.
-            Value::Array(items, _kind) => {
-                let mut result = Vec::new();
-                for item in items.iter() {
-                    match self.call_sub_value(block.clone(), vec![item.clone()], false) {
-                        Ok(v) => result.push(v),
-                        Err(e) if e.is_next() => continue,
-                        Err(e) => return Err(e),
-                    }
-                }
-                Ok(Value::array(result))
-            }
-            Value::Seq(items) => {
-                let mut result = Vec::new();
-                for item in items.iter() {
-                    match self.call_sub_value(block.clone(), vec![item.clone()], false) {
-                        Ok(v) => result.push(v),
-                        Err(e) if e.is_next() => continue,
-                        Err(e) => return Err(e),
-                    }
-                }
-                Ok(Value::array(result))
-            }
-            // Single value: apply the block directly
-            _ => self.call_sub_value(block.clone(), vec![target.clone()], false),
-        }
-    }
-
-    /// Apply duckmap to a single element: try the block, on failure descend.
-    fn duckmap_element(&mut self, block: &Value, value: &Value) -> Result<Value, RuntimeError> {
-        // Try to call the block with this value
-        match self.call_sub_value(block.clone(), vec![value.clone()], false) {
-            Ok(result) => Ok(result),
-            Err(e) if e.is_next() || e.is_last() || e.is_redo() => {
-                // Propagate loop control signals (next, last, redo)
-                Err(e)
-            }
-            Err(_) => {
-                // Block rejected the value (type mismatch, etc.)
-                // Try to descend into iterable structures
-                match value {
-                    Value::Array(items, kind) => {
-                        let mut result = Vec::new();
-                        for item in items.iter() {
-                            result.push(self.duckmap_element(block, item)?);
-                        }
-                        if kind.is_real_array() {
-                            Ok(Value::real_array(result))
-                        } else {
-                            Ok(Value::array(result))
-                        }
-                    }
-                    Value::Seq(items) => {
-                        let mut result = Vec::new();
-                        for item in items.iter() {
-                            result.push(self.duckmap_element(block, item)?);
-                        }
-                        Ok(Value::Seq(std::sync::Arc::new(result)))
-                    }
-                    Value::Hash(map) => {
-                        let mut result = std::collections::HashMap::new();
-                        for (k, v) in map.iter() {
-                            result.insert(k.clone(), self.duckmap_element(block, v)?);
-                        }
-                        Ok(Value::Hash(Value::hash_arc(result)))
-                    }
-                    // Not iterable — return unchanged
-                    _ => Ok(value.clone()),
-                }
-            }
-        }
     }
 }
