+++ src/runtime/mod.rs
@@ -135,6 +135,12 @@ mod builtins_atomic_cas;
 mod builtins_atomic_shared;
 mod builtins_coerce;
 pub(crate) mod builtins_collection;
+mod builtins_collection_classify;
+mod builtins_collection_deepmap;
+mod builtins_collection_extrema;
+mod builtins_collection_listops;
+mod builtins_collection_mapgrep;
+mod builtins_collection_minmax;
 mod builtins_control_flow;
 mod builtins_dispatch_next;
 mod builtins_eval_misc;
